Changes made after tests with Pi Camera
The Camera module didn't like storing the pictures in any directory
other than root. Changed it so that pictures are initially save in root,
uploaded via drop box and then moved to a specific directory.

The allows for a user to be able to move the files to media of storage
that could be set up as a NAS and still have a copy sent to their
Dropbox.

Plans for next interation are to introduce an option for a user to
disable the Dropbox uploading function.